TeamList: break out rows to its own component
